#02-Boas praticas em validações e implementações

+++ Controllers/ProdutoController.cs
@@ -2,6 +2,7 @@
 using Microsoft.EntityFrameworkCore;
 using Primeira_API.Data;
 using Primeira_API.Models;
+using Microsoft.AspNetCore.Http;
 
 namespace Primeira_API.Controllers
 {
@@ -17,34 +18,68 @@ namespace Primeira_API.Controllers
         }
 
         [HttpGet]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<ActionResult<IEnumerable<Produto>>> GetProdutos()
         {
+            if (_context.Produtos == null)
+            {
+                return NotFound();
+            }
+
             return await _context.Produtos.ToListAsync();
         }
 
         [HttpGet("{id:int}")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<ActionResult<Produto>> GetProduto(int id)
         {
+            if (_context.Produtos == null)
+            {
+                return NotFound();
+            }
+
             var produto = await _context.Produtos.FindAsync(id);
+
+            if (produto == null)
+            {
+                return NotFound();
+            }
 
             return produto;
         }
 
         [HttpPost]
+        [ProducesResponseType(StatusCodes.Status201Created)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesDefaultResponseType]
         public async Task<ActionResult<Produto>> PostProduto(Produto produto)
         {
+            if (_context.Produtos == null)
+            {
+                return Problem("Erro ao criar um produto, contate o suporte!");
+            }
+
             if (!ModelState.IsValid)
             {
                 return ValidationProblem(ModelState);
             }
 
             _context.Produtos.Add(produto);
+
             await _context.SaveChangesAsync();
 
             return CreatedAtAction(nameof(GetProduto), new { id = produto.Id }, produto);
         }
 
         [HttpPut("{id:int}")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<IActionResult> PutProduto(int id, Produto produto)
         {
             if (id != produto.Id) return BadRequest();
@@ -54,23 +89,57 @@ namespace Primeira_API.Controllers
                 return ValidationProblem(ModelState);
             }
 
-            _context.Produtos.Update(produto);
+            _context.Entry(produto).State = EntityState.Modified;
+
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                if (!ProdutoExist(id))
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    throw;
+                }                    
+            }
+
             await _context.SaveChangesAsync();
 
             return NoContent();
         }
 
         [HttpDelete("{id:int}")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<IActionResult> DeleteProduto(int id)
         {
+            if (_context.Produtos == null)
+            {
+                return NotFound();
+            }
+
             var produto = await _context.Produtos.FindAsync(id);
 
+            if (produto == null)
+            {
+                return NotFound();
+            }
+
             _context.Produtos.Remove(produto);
+
             await _context.SaveChangesAsync();
 
             return NoContent();
         }
         
-
+        private bool ProdutoExist(int id)
+        {
+            return (_context.Produtos?.Any(p => p.Id == id)).GetValueOrDefault();
+        }
     }
 }
